refactor: remove legacy registries support and update related tests
